PR: Mise à jour de Python de 3.7 à 3.8

+++ sardes/widgets/waitingspinner.py
@@ -103,10 +103,11 @@ class WaitingSpinner(QWidget):
                 sf = 1
 
             painter.setBrush(color)
-            rect = QRect(0, -self._lineWidth/2,
-                         self._lineLength * sf, self._lineWidth * sf)
+            rect = QRect(
+                0, int(-self._lineWidth/2),
+                int(self._lineLength * sf), int(self._lineWidth * sf))
             painter.drawRoundedRect(
-                    rect, self._roundness, self._roundness, Qt.RelativeSize)
+                rect, self._roundness, self._roundness, Qt.RelativeSize)
             painter.restore()
 
     def start(self):
@@ -216,12 +217,12 @@ class WaitingSpinner(QWidget):
         self.update()
 
     def updateSize(self):
-        size = (self._innerRadius + self._lineLength) * 2
+        size = int((self._innerRadius + self._lineLength) * 2)
         self.setFixedSize(size, size)
 
     def updateTimer(self):
-        self._timer.setInterval(
-                1000/(self._numberOfLines*self._revolutionsPerSecond))
+        self._timer.setInterval(int(
+                1000 / (self._numberOfLines*self._revolutionsPerSecond)))
 
     def updatePosition(self):
         if self.parentWidget() and self._centerOnParent:
